+++ packages/nuxt-locokit/src/pages/auth/AlreadyAuthenticated.vue
@@ -1,40 +1,39 @@
 <template>
-  <WithBackground>
-    <PrimeCard class="p-2 min-h-12">
-      <template #title>
+  <WithHeader>
+    <div
+      class="max-w-lg lg:h-full lg:m-auto mt-8 pb-4 px-4 lg:px-0 flex flex-col justify-center"
+    >
+      <div class="mb-8">
         <h1 class="text-center">
           {{ $t('pages.alreadyAuthenticated.title') }}
         </h1>
-      </template>
-      <template #content>
-        <div>
-          <p class="mb-4">
-            {{ $t('pages.alreadyAuthenticated.message') }}
-          </p>
-          <div class="flex justify-around">
-            <NuxtLink :to="{ name: ROUTES_NAMES.WORKSPACE.HOME }">
-              <PrimeButton
-                :label="$t('pages.alreadyAuthenticated.workspaceButton')"
-                icon="bi bi-person-workspace"
-              />
-            </NuxtLink>
+      </div>
+      <div>
+        <p class="mb-8">
+          {{ $t('pages.alreadyAuthenticated.message') }}
+        </p>
+        <div class="flex justify-around">
+          <NuxtLink :to="{ name: ROUTES_NAMES.WORKSPACE.HOME }">
             <PrimeButton
-              :label="$t('pages.alreadyAuthenticated.logoutButton')"
-              icon="bi bi-door-open-fill"
-              @click="logout"
+              :label="$t('pages.alreadyAuthenticated.workspaceButton')"
+              icon="bi bi-person-workspace"
             />
-          </div>
+          </NuxtLink>
+          <PrimeButton
+            :label="$t('pages.alreadyAuthenticated.logoutButton')"
+            icon="bi bi-door-open-fill"
+            @click="logout"
+          />
         </div>
-      </template>
-    </PrimeCard>
-  </WithBackground>
+      </div>
+    </div>
+  </WithHeader>
 </template>
 
 <script setup lang="ts">
-import PrimeCard from 'primevue/card'
 import PrimeButton from 'primevue/button'
 import { useI18n } from 'vue-i18n'
-import WithBackground from '../../layouts/WithBackground.vue'
+import WithHeader from '../../layouts/WithHeader.vue'
 import { useStoreAuth } from '../../stores/auth'
 import { ROUTES_NAMES } from '../../paths'
 import { useHead, useRouter } from '#imports'
