+++ packages/nuxt-locokit/src/pages/auth/LostPassword.vue
@@ -1,12 +1,14 @@
 <template>
-  <WithBackground>
-    <PrimeCard class="p-2">
-      <template #title>
-        <h1 class="text-center mb-4">
+  <WithHeader>
+    <div
+      class="max-w-lg lg:h-full lg:m-auto mt-8 pb-4 px-4 lg:px-0 flex flex-col justify-center"
+    >
+      <div class="mb-8">
+        <h1 class="text-center">
           {{ $t('pages.lostPassword.alternativeTitle') }}
         </h1>
-      </template>
-      <template #content>
+      </div>
+      <div>
         <div v-if="!formSentAndValid">
           <p class="mb-4">{{ $t('pages.lostPassword.description') }}</p>
           <LostPasswordForm
@@ -34,17 +36,16 @@
             <p>{{ $t('pages.lostPassword.homeLink') }}</p>
           </NuxtLink>
         </div>
-      </template>
-    </PrimeCard>
-  </WithBackground>
+      </div>
+    </div>
+  </WithHeader>
 </template>
 
 <script setup lang="ts">
-import PrimeCard from 'primevue/card'
 import { LostPasswordForm } from '@locokit/designsystem'
 import { storeToRefs } from 'pinia'
 import { useI18n } from 'vue-i18n'
-import WithBackground from '../../layouts/WithBackground.vue'
+import WithHeader from '../../layouts/WithHeader.vue'
 import { useStoreAuth } from '../../stores/auth'
 import { ROUTES_NAMES, ROUTES_PATH } from '../../paths'
 import { definePageMeta, ref, useHead } from '#imports'
