+++ packages/nuxt-locokit/src/pages/auth/VerifySignup.vue
@@ -1,10 +1,12 @@
 <template>
-  <WithBackground>
-    <PrimeCard class="flex-grow p-2 max-w-2xl">
-      <template #title>
-        <h1 class="text-center mb-4">{{ $t('pages.verifySignup.title') }}</h1>
-      </template>
-      <template #content>
+  <WithHeader>
+    <div
+      class="max-w-lg lg:h-full lg:m-auto mt-8 pb-4 px-4 lg:px-0 flex flex-col justify-center"
+    >
+      <div class="mb-8">
+        <h1 class="text-center">{{ $t('pages.verifySignup.title') }}</h1>
+      </div>
+      <div>
         <div v-if="!formSentAndValid">
           <p class="mb-4">{{ $t('pages.verifySignup.description') }}</p>
           <PasswordForm
@@ -32,17 +34,16 @@
             {{ $t('pages.resetPassword.homeLink') }}
           </NuxtLink>
         </div>
-      </template>
-    </PrimeCard>
-  </WithBackground>
+      </div>
+    </div>
+  </WithHeader>
 </template>
 
 <script setup lang="ts">
-import PrimeCard from 'primevue/card'
 import { PasswordForm } from '@locokit/designsystem'
 import { storeToRefs } from 'pinia'
 import { useI18n } from 'vue-i18n'
-import WithBackground from '../../layouts/WithBackground.vue'
+import WithHeader from '../../layouts/WithHeader.vue'
 import { ROUTES_NAMES } from '../../paths'
 import { useStoreAuth } from '../../stores/auth'
 import { definePageMeta, ref, useHead, useRoute } from '#imports'
